requirements.txt
aspose-words
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
